requirements.txt
astroid==1.6.2
autopep8==1.3.4
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
dj-database-url==0.5.0
Django==2.0.3
django-colorfield==0.1.14
django-cors-headers==2.2.0
django-extensions==2.0.7
django-simple-email-confirmation==0.23
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
idna==2.6
isort==4.3.4
lazy-object-proxy==1
Pillow==5.1.0
psycopg2==2.7.4
pycodestyle==2.3.1
PyJWT==1.6.1
pylint==1.8.3
python-dotenv==0.8.2
pytz==2018.3
requests==2.18.4
six==1.11.0
urllib3==1.22
whitenoise==3.3.1
wrapt==1.10.11
